Use BlankObject + method_missing instead of DelegateClass
DelegateClass + ActiveRecord::Base was not working with method delegation, and
after a lot of debug I still couldn't find the real reason behind it. I will
be investigating this further.

Besides that, this will allow us to easily integrate with other ORMs, not only
ActiveRecord.

All tests are now green on both 1.9.2 and 1.8.7.