Change localized behavior
Remove the possibility to give a list of attributes that should be localized
to the localized method. Instead, it will quack more like ActiveRecord, you can
give the attributes hash and extra options, and these attributes will be parsed
and assigned to the record.

This makes the usage much more friendly than the old behavior.